omniagent/backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-community
faiss-cpu
python-dotenv
google-generativeai
pydantic